requirements.txt
beautifulsoup4==4.1.3
distribute==0.6.31
mutagen==1.21
python-magic==0.4.3
requests==1.1.0
wsgiref==0.1.2
